+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 2.3.4
+    [#132](https://github.com/singer-io/tap-bing-ads/pull/132)
+  * Bump bingads from 13.0.11.1 to 13.0.28
+  * Fix a bug in complex_type parsing
+  * Update error handling to raise human readable message if it exists
+  * Update tests to use env vars for test account info
+
 # 2.3.3
   * Bump requests to 2.33.0 for security updates [#127](https://github.com/singer-io/tap-bing-ads/pull/127)